+++ src/api.py
@@ -4,6 +4,7 @@
 from fastapi import Request, Header, Response
 from fastapi.responses import JSONResponse
 from starlette.exceptions import HTTPException as StarletteHTTPException
+from fastapi.exceptions import RequestValidationError
 from datetime import datetime, timedelta
 import json, os, sys, time
 
@@ -74,40 +75,44 @@ async def languages():
     t = sorted(t)
     return {"error": False, "response": {"company": config.language, "supported": t}}
 
-cnt500 = []
-
 # thread to reload service
 def reload():
     time.sleep(1)
     os.system(f"./launcher hub restart {config.abbr} &")
 
-# error handler to uniform error response
+# error handler to format error response
 @app.exception_handler(StarletteHTTPException)
 async def errorHandler(request: Request, exc: StarletteHTTPException):
     return JSONResponse({"error": True, "descriptor": exc.detail}, status_code = exc.status_code)
 
+@app.exception_handler(RequestValidationError)
+async def error422Handler(request: Request, exc: RequestValidationError):
+    return JSONResponse({"error": True, "descriptor": "Unprocessable Entity"}, status_code = 422)
+
+err500 = []
 @app.exception_handler(500)
 async def error500Handler(request: Request, exc: Exception):
-    global cnt500
-    if not -1 in cnt500:
-        cnt500.append(time.time())
-        for i in range(len(cnt500)):
-            if cnt500[i] <= time.time() - 300:
-                cnt500.remove(cnt500[i])
-        if len(cnt500) >= 5:
+    global err500
+    if not -1 in err500:
+        err500.append(time.time())
+        err500[:] = [i for i in err500 if i > time.time() - 300]
+        if len(err500) >= 5:
             try:
                 requests.post(config.webhook_audit, data=json.dumps({"embeds": [{"title": "Attention Required", "description": "System detected too many `500 Internal Server Error`. API will restart automatically.", "color": config.intcolor, "footer": {"text": "System"}, "timestamp": str(datetime.now())}]}), headers={"Content-Type": "application/json"})
             except:
                 pass
             threading.Thread(target=reload).start()
-            cnt500.append(-1)
+            err500.append(-1)
 
-    return JSONResponse({"error": True, "descriptor": "Internal Server Error"}, status_code = 500)
+    if str(exc).lower().find("mysql") != -1: # probably lost connection
+        return JSONResponse({"error": True, "descriptor": "Service Unavailable"}, status_code = 503)
+    else:
+        return JSONResponse({"error": True, "descriptor": "Internal Server Error"}, status_code = 500)
         
 @app.on_event("startup")
-async def startUpEvent():
+async def startupEvent():
     await aiosql.create_pool()
 
 @app.on_event("shutdown")
 async def shutdownEvent():
-    await aiosql.shutdown()
+    await aiosql.close_pool()